Various wording updates.
I'm not married to any of these changes, if you don't agree with them :).

I shuffled the order of the Web SDK implementation methods to include Tags first, because, according to the AEP docs, that is the recommended method. I also added "(Recommended)" here. 

I also pulled the other "(Recommended)" statements out of the info icons to make them more prominent. And I removed the "(Not recommended), because the aforementioned change makes them kind of redundant.